requirements.txt
Django==5.2.4
psycopg2-binary>=2.9
django-environ>=0.11
dj-database-url>=2.2
Pillow>=10.0
requests>=2.31
django-cors-headers>=4.3
sentry-sdk>=2.0
django-widget-tweaks>=1.5
django-extensions>=3.2
django-formtools>=2.5
django-crispy-forms>=2.3
crispy-bootstrap5>=2024.2
django-axes>=6.5
djangorestframework>=3.15
drf-yasg>=1.21
whitenoise>=6.7
django-allauth>=0.63
django-csp>=3.8
django-otp>=1.4
django-two-factor-auth>=1.15
qrcode>=7.4
PyJWT>=2.9
google-auth>=2.34
google-auth-oauthlib>=1.2
cryptography>=43.0.0
